LayoutTests/http/tests/loading/empty-urls-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
PASS: img onerror fired
PASS: input type=image onerror fired
PASS: script onerror fired
PASS: video src onerror fired
PASS: audio onerror fired
PASS: source onerror fired
      
